Initial Setup
Finally got the repo to work properly. This is a commit of my work over
the past two days. I have set up the basic html for the tutorial pages,
the intro, and the extra help. I have also written php includes for the
header, the footer, and the practice problem form. The practice
problems do not work yet, they just display the question.