Tester and Nav Additions
I added the ability to detect improper regexes
 and tell the user that there is an error. I also fixed a small error
in the tester that affected the first match. Finally, I added an
unordered list that has all of the nav links in it that lead to all of
the pages we have made already. Empty links are there for the ones that
have yet to be made.