Added Answer Checking for Practice Problems
I had to change some of the files to make them work with my answer
checking script, so that’s what most of the changes are. Then, of
course, I wrote a script that validates that the user entered an answer
and that their answer is right for the practice problems. I also added
the beginning of the dictionary file.